requirements.txt
bscscan-python >= 2.0.0
cryptocompare